Changed and add one api call

+++ routes/api.php
@@ -27,30 +27,48 @@ Route::get('/getroom', function (Request $request) {
     return response(json_encode($room))->header('Content-Type', 'application/json');
 });
 
-Route::post('/setrooms', function (Request $request) {
-    $numberOfRooms = (int) $request->rooms;
+Route::post('/addroom', function (Request $request) {
+    $arduinoId = $request->arduino_id;
     $userId = $request->user_id;
 
     $user = User::where('id', $userId)->first();
 
     if ($user) {
-        $rooms = Room::all();
-        foreach ($rooms as $room){
-            $room->delete();
-        }
-        for ($i = 0; $i < $numberOfRooms; $i++) {
             $newRoom = new Room();
-            $newRoom->name = 'Soba' . $i;
+            $newRoom->name = 'Soba' . $userId . $arduinoId;
             $newRoom->start_time = new DateTime('00:00:00');
             $newRoom->end_time = new DateTime('23:45:00');
             $newRoom->automatic = false;
             $newRoom->brightness = 255;
-            $newRoom->arduino_id = $i;
+            $newRoom->arduino_id = $arduinoId;
             $newRoom->user_id = $user->id;
 
             $newRoom->save();
+
+        $responseArray = array('status' => 1, 'response' => 'Room added.');
+    }else{
+        $responseArray = array('status' => 0, 'response' => 'Can\'t find user with that id.');
+    }
+
+
+    return response(json_encode($responseArray))->header('Content-Type', 'application/json');
+});
+
+Route::post('/deleteroom', function (Request $request) {
+    $arduinoId =  $request->arduino_id;
+    $userId = $request->user_id;
+
+    $user = User::where('id', $userId)->first();
+    if ($user) {
+        $room = Room::where(array('user_id' => $userId,'arduino_id' => $arduinoId ))->first();
+
+        if ($room) {
+            $room->delete();
+
+            $responseArray = array('status' => 1, 'response' => 'Room deleted.');
+        }else{
+            $responseArray = array('status' => 0, 'response' => 'Can\'t find room.');
         }
-        $responseArray = array('status' => 1, 'response' => 'Rooms added.');
     }else{
         $responseArray = array('status' => 0, 'response' => 'Can\'t find user with that id.');
     }
